Sanity check free timing on failures
Right now when it works in CI the free happens ~1.3ms after the call to
the delay trait (which makes sense given the 1ms delay in Rust before it
aborts). When it doesn't work though the free doesn't happen until ~.6ms
after the delay call fully finishes, so 100.5ms after the call to the
delay trait. It seems like a bug where the FreeImpl isn't being called
from Rust in time, but I want to see if bumping the delay to 200ms
results in 1.3/200.5, or 1.3/100.5 still